requirements.txt
Django==4.1.4
python-decouple==3.7
django-bootstrap4==22.3
Pillow==9.4.0
django-cleanup==6.0.0
easy-thumbnails==2.8.5
django-cors-headers==3.13.0
djangorestframework==3.14.0